// src/components/textField/types.d.ts
export type ColorType = string | {
    default?: string;
    focus?: string;
    error?: string;
    disabled?: string;
    readonly?: string;
};
export declare enum ValidationMessagePosition {
    TOP = "top",
    BOTTOM = "bottom"
}
export declare enum Presets {
    DEFAULT = "default",
    UNDERLINE = "underline",
    OUTLINE = "outline"
}
export type ValidationMessagePositionType = `${ValidationMessagePosition}` | ValidationMessagePosition;
export type Validator = ((value?: string) => boolean) | keyof typeof formValidators;
// ...
export interface MandatoryIndication {
    /**
     * Whether to show a mandatory field indication.
     */
    showMandatoryIndication?: boolean;
}
// ...
export interface FloatingPlaceholderProps extends MandatoryIndication {
    /**
     * The placeholder for the field
     */
    placeholder?: string;
    /**
     * The floating placeholder color
     */
    floatingPlaceholderColor?: ColorType;
    /**
     * Custom style to pass to the floating placeholder
     */
    floatingPlaceholderStyle?: StyleProp<TextStyle>;
    /**
     * Should placeholder float on focus or when start typing
     */
    floatOnFocus?: boolean;
    validationMessagePosition?: ValidationMessagePositionType;
    extraOffset?: number;
    defaultValue?: TextInputProps['defaultValue'];
    testID: string;
}
// ...
export type FieldContextType = {
    value?: string;
    isFocused: boolean;
    hasValue: boolean;
    isValid: boolean;
    failingValidatorIndex?: number;
    disabled: boolean;
    readonly: boolean;
    validateField: () => void;
    checkValidity: () => boolean;
    isMandatory: boolean;
};